Add files via upload
This project involved the creation and refinement of a convolutional neural network (CNN) model, leveraging the ResNet-18 architecture, aimed at automating the process of cell counting in histopathology images. The primary goal was to develop a reliable and efficient AI tool that could assist pathologists and researchers in quantifying cell populations in biomedical images, a task critical for various diagnostic and research applications. Using the https://conic-challenge.grand-challenge.org/ dataset.